+++ hata/ext/slash/__init__.py
@@ -19,6 +19,7 @@ from .wrappers import *
 __all__ = (
     'P',
     'configure_parameter',
+    'Form',
     'set_permission',
     'setup_ext_slash',
     *application_command.__all__,
@@ -46,6 +47,9 @@ from .client_wrapper_extension import *
 set_permission = SlasherApplicationCommandPermissionOverwriteWrapper
 configure_parameter = SlasherApplicationCommandParameterConfigurerWrapper
 P = SlashParameter
+
+from ...discord.interaction import InteractionForm as Form
+
 
 def setup_ext_slash(client, **kwargs):
     """